Added dom version when requiring css is not possible
You can now do require('bespoke-notes/dom');
